+++ database/migrations/2022_09_04_202418_create_category_filter_table.php
@@ -14,8 +14,8 @@ return new class extends Migration
     public function up()
     {
         Schema::create('category_filter', function (Blueprint $table) {
-            $table->integer('category_id');
-            $table->integer('filter_id');
+            $table->unsignedBigInteger('category_id');
+            $table->unsignedBigInteger('filter_id');
             
         });
     }
