+++ database/migrations/2022_09_10_204522_create_user_advertisment_table.php
@@ -14,8 +14,8 @@ return new class extends Migration
     public function up()
     {
         Schema::create('user_advertisment', function (Blueprint $table) {
-            $table->integer('user_id');
-            $table->integer('advertisment_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('advertisment_id');
         
         });
     }
